update: move common types into dedicated file
