+++ src/types/index.ts
@@ -1,0 +1,10 @@
+export interface RankingI {
+  score: number;
+  nickname: string;
+}
+
+export interface RankingResponse {
+  success: boolean;
+  rankings?: RankingI[];
+  user?: { score: number; ranking: number };
+}
